pyproject.toml
[tool.poetry]
name = "miiflow-agent"
version = "1.0.0"
description = "A lightweight Python SDK for building AI agents with unified LLM providers, ReAct orchestration, and tool calling"
authors = [ "Yaozhong Kang <yaozhong@miiflow.ai>" , "Debjyoti Ray <dev@miiflow.ai>" ]
maintainers = [ "Yaozhong Kang <yaozhong@miiflow.ai>" ]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/Miiflow/miiflow-agent"
repository = "https://github.com/Miiflow/miiflow-agent"
documentation = "https://github.com/Miiflow/miiflow-agent#readme"
keywords = ["agent", "llm", "openai", "anthropic", "react", "orchestration", "tools"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
openai = "^1.50.0"
anthropic = "^0.74.0"
pydantic = "^2.0.0"
tiktoken = "^0.8.0"
tenacity = "^9.0.0"
aiohttp = "^3.9.0"
httpx = "^0.28.0"
structlog = "^25.3.0"
typing-extensions = "^4.8.0"

# Optional dependencies for specific providers
groq = { version = "^0.13.0", optional = true }
google-generativeai = { version = "^0.8.0", optional = true }
mistralai = { version = "^1.0.0", optional = true }

# Optional multimedia dependencies
PyMuPDF = { version = "^1.23.0", optional = true }

# Optional example dependencies
yfinance = { version = "^0.2.40", optional = true }

# Optional MCP (Model Context Protocol) support
mcp = { version = "^1.0.0", optional = true, python = ">=3.10" }

# Optional AG-UI (Agent-User Interaction Protocol) support
ag-ui-protocol = { version = "^0.1.10", optional = true }

# Optional observability dependencies
opentelemetry-api = { version = "^1.27.0", optional = true }
opentelemetry-sdk = { version = "^1.27.0", optional = true }
opentelemetry-exporter-otlp = { version = "^1.27.0", optional = true }
prometheus-client = { version = "^0.21.0", optional = true }
arize-phoenix = { version = "^4.29.0", optional = true }

# OpenInference auto-instrumentation
openinference-instrumentation-openai = { version = "^0.1.12", optional = true }
openinference-instrumentation-anthropic = { version = "^0.1.5", optional = true }
openinference-instrumentation-google-genai = { version = "^0.1.0", optional = true }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.25.0"
pytest-cov = "^6.0.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.11.0"
pre-commit = "^4.0.0"
python-dotenv = "^1.2.1"

[tool.poetry.extras]
groq = ["groq"]
google = ["google-generativeai"]
mistral = ["mistralai"]
pdf = ["PyMuPDF"]
multimedia = ["PyMuPDF"]
examples = ["yfinance"]
mcp = ["mcp"]
agui = ["ag-ui-protocol"]
observability = ["opentelemetry-api", "opentelemetry-sdk", "opentelemetry-exporter-otlp", "prometheus-client", "arize-phoenix", "openinference-instrumentation-openai", "openinference-instrumentation-anthropic", "openinference-instrumentation-google-genai"]
all = ["groq", "google-generativeai", "mistralai", "PyMuPDF", "yfinance", "mcp", "ag-ui-protocol", "opentelemetry-api", "opentelemetry-sdk", "opentelemetry-exporter-otlp", "prometheus-client", "arize-phoenix", "openinference-instrumentation-openai", "openinference-instrumentation-anthropic", "openinference-instrumentation-google-genai"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers"
]
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as async"
]
